refactor(standard report): adjust tests to refactored state
